fixes travis CI error [Werror=sign-compare]

+++ src/utils_str.c
@@ -1239,9 +1239,9 @@ gchar *utils_str_remove_accents (const gchar *text)
     gchar *ptr;
     gchar *tmp_str;
     gunichar ch;
-    gint i = 0;
+    glong i = 0;
     size_t nbre_bytes;
-    size_t nbre_chars;
+    glong nbre_chars;
 
 	if (!text)
 		return NULL;
